rename lib/repositories.js to lib/repos.js and remove old profile tests
